requirements.txt
numpy>=1.24
scipy>=1.10
torch>=2.0
matplotlib>=3.7
pyyaml>=6.0
pytest>=7.0
